Add files via upload
In this commit I have filled the project with the new files I have created since the last update, as the total project is too big. 
I created the Register and Login buttons, validated them, the user automatically enters the database when creating an account and can log in and log out at any time.
I made buttons for manage, add, edit and delete topics (for this one we also made their validation) and for users and posts.