Create README.md
In this commit I have filled the project with the new files I have created since the last update, as the total project is too big. 
I made login and register buttons that we validated, where the user can create an account that will go into the database, log in to it and also log out. The user cannot make an account with the same email twice.
I made buttons for manage, add, edit and delete topics (for this one we also made their validation) and for users and posts.